requirements.txt
# A0 SkillArchitect dependencies
# (Python logic primarily delegates to Node.js)
markdown
PyYAML
